new router and some updates

+++ routers/user.js
@@ -2,21 +2,20 @@ const express = require('express');
 const models = require('../models');
 const router = express.Router();
 const { StatusCodes } = require('http-status-codes');
+const tokenGen = require('../utils/token');
 
 router.post('/user', async (req, res) => {
   try {
     const user = await models.User.create({ ...req.body });
-    return res.status(StatusCodes.CREATED).send(user);
+    const token = tokenGen({ userId: user.id });
+    return res.status(StatusCodes.CREATED).send({ token });
   } catch (e) {
-    res.status(401).send();
+    res.status(StatusCodes.INTERNAL_SERVER_ERROR).send(e);
   }
 });
 
 router.get('/user/:id', async (req, res) => {
   const { id: _id } = req.params;
-  // const { params: {
-  //   id: _id
-  // }} = req;
   try {
     const user = await models.User.findOne({ where: { id: _id } });
     res.status(StatusCodes.OK).send(user);
@@ -39,10 +38,12 @@ router.put('/user/:id', async (req, res) => {
 router.delete('/user/:id', async (req, res) => {
   const _id = req.params.id;
   try {
+    const userId = await models.User.findOne({ where: { id: _id } });
+    if (!userId) return res.status(StatusCodes.NOT_FOUND).send();
     await models.User.destroy({ where: { id: _id } });
     res.status(StatusCodes.OK).send('deleted');
   } catch (e) {
-    res.status(StatusCodes.NOT_FOUND).send();
+    return res.status(StatusCodes.NOT_FOUND).send();
   }
 });
